Minor enhancements
